Sentiment Analysis and Text Mining/requirements.txt
matplotlib.pyplot
numpy
sys
networkx